[fix] board 코드 리뷰 반영(userEmail, 검증 로직 추가, board삭제 분기 처리)

+++ src/main/java/teamdevhub/devhub/core/board/application/CommentService.java
@@ -20,7 +20,7 @@ import teamdevhub.devhub.shared.enums.ErrorCode;
 @Transactional
 @RequiredArgsConstructor
 public class CommentService implements CommentUseCase {
-	
+
 	private final IdentifierProvider identifierProvider;
 	private final CommentRepository commentRepository;
 	private final UserRepository userRepository;
@@ -32,30 +32,45 @@ public class CommentService implements CommentUseCase {
         commentList.forEach(comment ->
         	comment.fillUserName(commentUserNames.getOrDefault(comment.getUserGuid(), ""))
         		);
-        
+
         return commentList;
 	}
-	
+
 	@Override
 	public void createComment(CreateCommentCommand createCommentCommand) {
 		String commentGuid = identifierProvider.generateIdentifier();
 		Comment comment = Comment.createComment(createCommentCommand, commentGuid);
 		commentRepository.save(comment);
 	}
-	
+
 	@Override
-	public void deleteComment(String boardGuid, String commentGuid) {
+	public void deleteComment(String boardGuid, String commentGuid, String userGuid) {
+		Comment comment = commentRepository.findByCommentGuid(commentGuid);
+		validateCommentBoard(comment, boardGuid);
+		validateOwner(comment.getUserGuid(), userGuid);
 		commentRepository.deleteByBoardGuidAndCommentGuid(boardGuid, commentGuid);
 	}
-	
+
 	@Override
 	public void updateComment(UpdateCommentCommand updateCommentCommand) {
 		Comment comment = commentRepository.findByCommentGuid(updateCommentCommand.commentGuid());
-		if (!comment.getBoardGuid().equals(updateCommentCommand.boardGuid())) {
-	        throw BusinessRuleException.of(ErrorCode.READ_FAIL);
-	    }
-		
+
+		validateCommentBoard(comment, updateCommentCommand.boardGuid());
+		validateOwner(comment.getUserGuid(), updateCommentCommand.userGuid());
+
 		comment.updateContent(updateCommentCommand.content());
 		commentRepository.updateComment(comment);
+	}
+
+	private void validateCommentBoard(Comment comment, String boardGuid) {
+		if (!comment.getBoardGuid().equals(boardGuid)) {
+	        throw BusinessRuleException.of(ErrorCode.READ_FAIL);
+	    }
+	}
+
+	private void validateOwner(String ownerGuid, String requestUserGuid) {
+		if (!ownerGuid.equals(requestUserGuid)) {
+			throw BusinessRuleException.of(ErrorCode.AUTH_INVALID);
+		}
 	}
 }
